// SOK/frontend/src/model/IRegisterUser.ts
export interface IRegisterUser {
  readonly firstName: string;
  readonly lastName: string;
  readonly secondName: string;
  readonly email: string;
  readonly userName: string;
  readonly password: string;
  readonly addrIndex: string;
  readonly addrCity: string;
  readonly addrSreet: string;
  readonly addrHouse: string;
  readonly addrStructure?: string;
  readonly addrApart?: string;
}